fix: address codex review for cross-chapter offline flow
