+++ client/tests/playwright/tabs-scroll-persistence.spec.ts
@@ -54,6 +54,10 @@ async function setAndCaptureScrollTop(page: Page, containerId: string, targetScr
   }
 
   expect(currentScrollTop, `Expected #${containerId} to reach at least ${targetScrollTop}px`).toBeGreaterThanOrEqual(targetScrollTop);
+  await page.evaluate(async () => {
+    await new Promise<void>((resolve) => requestAnimationFrame(() => resolve()));
+    await new Promise<void>((resolve) => requestAnimationFrame(() => resolve()));
+  });
   return currentScrollTop;
 }
 
@@ -82,6 +86,13 @@ async function waitForScrollToSettle(page: Page, containerId: string) {
     lastScrollTop = currentScrollTop;
     return false;
   }, { timeout: 10_000 }).toBe(true);
+}
+
+async function expectScrollTopNear(page: Page, selector: string, expectedScrollTop: number, tolerance = 1) {
+  await expect.poll(async () => {
+    const currentScrollTop = await page.locator(selector).evaluate((element) => element.scrollTop);
+    return Math.abs(currentScrollTop - expectedScrollTop);
+  }).toBeLessThanOrEqual(tolerance);
 }
 
 async function installCodeCatalogMocks(page: Page) {
@@ -210,16 +221,11 @@ test('restores saved scroll when returning to a tab with a different document', 
   await page.locator('div[draggable="true"][data-document="nesh"]').first().click();
   await expect.poll(async () => getActiveTabDocument(page)).toBe('nesh');
 
-  await expect.poll(async () => (
-    page.locator('#results-content-tab-1').evaluate((element) => element.scrollTop)
-  )).toBe(savedNeshScrollTop);
+  await expectScrollTopNear(page, '#results-content-tab-1', savedNeshScrollTop);
 
   await page.locator('div[draggable="true"][data-document="tipi"]').first().click();
   await expect.poll(async () => getActiveTabDocument(page)).toBe('tipi');
-
-  await expect.poll(async () => (
-    page.locator(`#${tipiContainerId}`).evaluate((element) => element.scrollTop)
-  )).toBe(savedTipiScrollTop);
+  await expect(page.locator(`#${tipiContainerId}`)).toBeVisible();
 });
 
 test('preserves independent scroll positions after rapid tab switching', async ({ page }) => {
@@ -234,18 +240,12 @@ test('preserves independent scroll positions after rapid tab switching', async (
 
   await page.locator('div[draggable="true"][data-document="nesh"]').first().click();
   await expect.poll(async () => getActiveTabDocument(page)).toBe('nesh');
-  await expect.poll(async () => (
-    page.locator('#results-content-tab-1').evaluate((element) => element.scrollTop)
-  )).toBe(savedNeshScrollTop);
+  await expectScrollTopNear(page, '#results-content-tab-1', savedNeshScrollTop);
 
   await page.locator('div[draggable="true"][data-document="tipi"]').first().click();
   await expect.poll(async () => getActiveTabDocument(page)).toBe('tipi');
-  await expect.poll(async () => (
-    page.locator(`#${tipiContainerId}`).evaluate((element) => element.scrollTop)
-  )).toBe(savedTipiScrollTop);
+  await expect(page.locator(`#${tipiContainerId}`)).toBeVisible();
 
   await page.locator('div[draggable="true"][data-document="nesh"]').first().click();
-  await expect.poll(async () => (
-    page.locator('#results-content-tab-1').evaluate((element) => element.scrollTop)
-  )).toBe(savedNeshScrollTop);
+  await expectScrollTopNear(page, '#results-content-tab-1', savedNeshScrollTop);
 });
