+++ tests/integration/test_search_route_contracts.py
@@ -5,6 +5,8 @@ import pytest
 
 from backend.presentation.routes import search as search_route
 from backend.presentation.routes import tipi as tipi_route
+from backend.infrastructure.database import DatabaseAdapter
+from backend.infrastructure.repositories.chapter_repository import ChapterRepository
 from backend.services.nesh_service import NeshService
 from backend.services.tipi_service import TipiService
 from backend.server.app import app
@@ -206,11 +208,16 @@ def test_search_chapter_body_allows_anonymous_access(client, monkeypatch):
 
 
 def test_search_chapters_endpoint_returns_available_chapters(client, monkeypatch):
-    class _FakeDb:
-        async def get_all_chapters_list(self):
-            return ["01", "02", "84"]
-
-    monkeypatch.setattr(app.state, "db", _FakeDb(), raising=False)
+    monkeypatch.setattr(
+        DatabaseAdapter,
+        "get_all_chapters_list",
+        AsyncMock(return_value=["01", "02", "84"]),
+    )
+    monkeypatch.setattr(
+        ChapterRepository,
+        "get_all_nums",
+        AsyncMock(return_value=["01", "02", "84"]),
+    )
 
     response = client.get("/api/chapters")
